+++ packages/client-core/src/key_pair/index.ts
@@ -1,4 +1,3 @@
 export { KeyType } from './constants';
-export { KeyPair, Signature } from './key_pair';
-export { KeyPairEd25519 } from './key_pair_ed25519';
+export { KeyPair, KeyPairEd25519, Signature } from './key_pair';
 export { PublicKey } from './public_key';
